Always add password when creating LDAP connection to change it
Even if we use the ldap admin to bind, we still need the current
password to change the password. It is a bit confusing, but using the
ldap admin to bind means that it is used for searching and connecting to
the ldap. It is not used to actually modify the password, thus the
current password should always be set on the connection. I know that
sounds silly as well, but that's how this silly gem works.